* main:
  Fixed various UBSan warnings about working with NULL pointers
  Fixed misaligned memory access flagged by UBSan
  disable test that depend on ncpathcvt in cmake build w/o utilities
  Update RELEASENOTES
  Remove the execinfo capability
  fix memory leak
  ckp
  remove conflicts
  Fixed various UBSan warnings about invalid bit shifting
  Add fenceposting so DLL flags are only introduced wwhen we are compiling DLL-based shared libraries.
  Working to get the proposed change working with Visual studio
  CMake: Use helper libraries for nczarr tests
  Fix most float conversion warnings
  Update release notes
  Update internal tinyxml2 code to the latest version
  Update release notes
  Improve fetch performance of DAP4

+++ libhdf5/hdf5var.c
@@ -2398,7 +2398,7 @@ nc_set_var_chunk_cache_ints(int ncid, int varid, int size, int nelems,
         real_nelems = nelems;
 
     if (preemption >= 0)
-        real_preemption = preemption / 100.;
+        real_preemption = (float)(preemption / 100.);
 
     return NC4_HDF5_set_var_chunk_cache(ncid, varid, real_size, real_nelems,
                                         real_preemption);
